+++ src/Adapter/SQL/Repository.php
@@ -30,7 +30,7 @@ use Innmind\Immutable\{
  * @template T of object
  * @implements RepositoryInterface<T>
  */
-final class Repository implements RepositoryInterface, RepositoryInterface\MassRemoval
+final class Repository implements RepositoryInterface
 {
     private Connection $connection;
     /** @var Definition<T> */